#9: add text theme to device app storage
